Code_texture_generation/texture-voronoi-diagrams/requirements.txt
matplotlib
opencv-python == 4.6.0.66
Pillow == 8.4.0
scipy == 1.5.4
tqdm == 4.64.0
albumentations == 1.3.0
numpy==1.26.4
